requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bootstrap4==0.1.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-bootstrap4==2.3.1
django-crispy-forms==1.11.1
django-excel==0.0.10
django-import-export==2.5.0
et-xmlfile==1.0.1
jdcal==1.4.1
lml==0.1.0
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.6
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
PyMySQL==1.0.2
pytz==2021.1
PyYAML==5.4.1
soupsieve==2.2
sqlparse==0.4.1
tablib==0.12.1
texttable==1.6.3
unicodecsv==0.14.1
xlrd==2.0.1
xlwt==1.3.0
